# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TrackStar API",
    "description": "Personal Task & Habit Tracker API",
    "version": "1.0.0",
    "contact": {
      "name": "TrackStar Team",
      "email": "support@trackstar.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/habits/": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "Get all habits for the current user",
        "description": "Retrieve all habits belonging to the authenticated user.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Create a new habit",
        "description": "Create a new habit for the authenticated user.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Habit information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Daily Exercise"
                },
                "description": {
                  "type": "string",
                  "example": "Do 30 minutes of exercise daily"
                },
                "frequency": {
                  "type": "string",
                  "example": "daily"
                },
                "targetCount": {
                  "type": "number",
                  "example": 1
                },
                "category": {
                  "type": "string",
                  "example": "health"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/habits/{id}": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "Get a specific habit by ID",
        "description": "Retrieve details of a specific habit by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Habits"
        ],
        "summary": "Update an existing habit",
        "description": "Update details of an existing habit by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated habit information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Updated habit name"
                },
                "description": {
                  "type": "string",
                  "example": "Updated description"
                },
                "frequency": {
                  "type": "string",
                  "example": "weekly"
                },
                "isActive": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Habits"
        ],
        "summary": "Delete a habit",
        "description": "Delete a specific habit by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "User ID for authentication (temporary for testing)"
      }
    }
  }
}